refactor(AngleAttackControl): move GUI layout drawing into the class
